Components: Input-element-config: Rename and implement on EditMenu
